Refactor: Move traffic implementation(valhalla_add_predecited_traffic) to traffic_updater
