+++ src/mjolnir/CMakeLists.txt
@@ -43,6 +43,7 @@ set(sources
   hierarchybuilder.cc
   ingest_transit.cc
   landmarks.cc
+  traffic_updater.cc
   linkclassification.cc
   luatagtransform.cc
   node_expander.cc
